refactor: dto를 infra -> domain으로 위치 변경
